Saikatsune
Update 0.9.9-PRE-RELEASE change-logs:

Optimized code.
Added combat-logger.
Added logout-timer.
Removed unnecessary features.

Fixed every issue that was known on previously version and it is the last pre-release version of this plugin. It will be released as 'full-release' after tests and last bug-fixes.

Everyone has permission to use it and/or use code of it for theirselves.

Have fun with it and report any bugs on Discord by contacting 'Saikatsune#2329'.